Split unit tests out into separate files
This will help us manage the tests as we add more operations and
maintain the existing ones.
