+++ Assets/RedBlueGames/BulkRename/Tests/Editor/AddStringOpTests.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: aa7406fb936d845579a18d277e414445
+timeCreated: 1480087188
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
