+++ Assets/RedBlueGames/BulkRename/Tests/Editor/BulkRenamerTests.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c430997fefca94d41a19946e67b4237c
+timeCreated: 1480087188
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
