+++ Assets/RedBlueGames/BulkRename/Tests/Editor/ChangeCaseOpTests.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ffd61cbb1996e485596e3e518c9f5eab
+timeCreated: 1480087188
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
