+++ Assets/RedBlueGames/BulkRename/Tests/Editor/EnumerateOpTests.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d17de448d687b471ba0d743e4d02a3c7
+timeCreated: 1480087188
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
